Removed the dpkg options added in previous commit
It turns out that cron-apt does not handle the dpkg options properly, causing
dpkg to fail with "dpkg: error: need an action option". When the apt-get
command-line created by cron-apt is run manually no problems occur. In any case,
the previous change prevented all automated upgrades, so decided to switch back
to the old configuration which works correctly most of the time at least. When
the dpkg config file prompt prevents an upgrade the next time another stab is
taken to fix the issue cleanly.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>